app/requirements.txt
flask==2.3.3
werkzeug==2.3.7
python-keycloak==3.0.0
hvac==1.1.0
python-dotenv==0.19.0
requests==2.31.0 